% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_distance_to_closest.r
\name{average_distance_to_closest}
\alias{average_distance_to_closest}
\title{average_distance_to_closest(inst, percent_closest)}
\usage{
average_distance_to_closest(inst, percent_closest)
}
\arguments{
\item{inst}{todo list}

\item{percent_closest}{todo list}
}
\value{
todo list
}
\description{
average_distance_to_closest(inst, percent_closest)
}
