% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_dist_closest.r
\name{avg_dist_closest}
\alias{avg_dist_closest}
\title{avg_dist_closest()}
\usage{
avg_dist_closest(proxim, percent_closest = 1)
}
\arguments{
\item{proxim}{An object corresponding to the output of the \code{proxim_dist} function}

\item{percent_closest}{A value between 0 and 1 (by default) corresponding to the desired percent of indivisuals that will be taken in the computation of the distances}
}
\value{
A list of 3 matrix is returned:
\item{Davg}{The cost matrix which number of rows corresponds to the number of levels of the target variable in database A, and which number of columns corresponds to the number of levels of the target variable in database B.
This cost matrix can be interpreted as the cost of moving from one level of the target factor in the 1st database A to one level of the target factor in the 2nd database B.
Davg[P,Q] so refers to the distance between covariates vectors of individuals in database A having modality P and individuals in database B having modality Q}
\item{DindivA}{A matrix which number of rows corresponds to the number of rows of the 1st database A and the number of columns corresponds to the number of levels of the target variable in the 2nd database B.
DindivA[i,Q] refers to the average distance of the i^th individual of the 1st database and a given proportion of individuals (\code{percent_closest} set by the user) of the 2nd database having the modality Q as value of the target variable}
\item{DindivB}{A matrix which number of rows corresponds to the number of rows of the 2nd database B and the number of columns corresponds to the number of levels of the target variable in the 1st database A.
DindivB[k,P] refers to the average distance of the k^th individual of the 1st database and a given proportion of individuals (\code{percent_closest} set by the user) of the 2nd database having the modality P as value of the target variable}
}
\description{
\code{avg_dist_closest} is an intermediate function of the Optimal Transportation algorithm (OT) that especially computes average distances between each individual of 2 databases (A and B) and levels of target factors that synthetise a same information in 2 different encodings.
}
\details{
\code{avg_dist_closest} is directly implemented in the \code{OT} and \code{OT_JOINT} functions but can also be used separately.
Nevertheless, this function can not be used without the \code{\link{proxim_dist}} function because it uses the objects from the latter.
So do not use this function directly on your data, and, if needed, do not hesitate to consult the example provided for a better understanding.


Definition of the cost matrix
------------------------------

\code{avg_dist_closest} uses, in particular, the D distance matrix (that stores distances between individuals of A and B) from \code{\link{proxim_dist}} to produce 3 distinct matrices stored in a list.
This matrix is used to calculate average distances between individuals in A (resp. B) and individuals of B (resp. A) related to each level of Z (resp. Y) in B (resp. A). This information is stored in \code{DinvidA} and \code{DinvidB} respectively.

A cost matrix called \code{Davg} which dimensions correspond to the number of levels of Y (rows) and Z (columns), is also generated to calculate average distances between covariations of a given part of individuals (the percent closest neighbors)  with identical levels of Y in A and identical levels of Z in B.
Davg[i,j] corresponds to average distances between covariations of a given part of individuals (in A) which related values for Y are i, and another given part of  individuals (in B) which related values for Z are j.

Thus, this matrix can be seen as the ability for an individual to move from a given class of a target variable (Y) in the 1st database (A) to another given class of the corresponding target variable (Z) in the 2nd database (B).
}
\references{
Gares V, Dimeglio C, Guernec G, Fantin F, Lepage B, Korosok MR, savy N (2019). On the use of optimal transportation theory to recode variables and application to database merging. The International Journal of Biostatistics.
0, 20180106 (2019) | \url{https://doi.org/10.1515/ijb-2018-0106}
}
\seealso{
\code{\link{proxim_dist}}
}
\author{
Gregory Guernec, Valerie Gares, Jeremy Omer
\email{gregory.guernec@inserm.fr}
}
