% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTRECOD_1.r
\name{compare_lists}
\alias{compare_lists}
\title{compare_lists()
   Function which compare the elements of two lists of same lengths
   The order of entry of the 2 lists have no importance}
\usage{
compare_lists(listA, listB)
}
\arguments{
\item{listA}{A first list}

\item{listB}{A second list}
}
\value{
Value A boolean vector (of same length as the two lists) 
        which ith answer is TRUE if the ith element is different 
        between the 2 lists, or FALSE otherwise
}
\description{
compare_lists()
   Function which compare the elements of two lists of same lengths
   The order of entry of the 2 lists have no importance
}
\examples{
data1 = data.frame(Gender = rep(c("m","f"),5),Age = stats::rnorm(5,20,4))
data2 = data.frame(Gender = rep(c("m","f"),5),Age = stats::rnorm(5,21,5))

list1 = list(A = 1:4, B = as.factor(c("A","B","C")), C = matrix(1:6,ncol = 3))
list2 = list(A = 1:4, B = as.factor(c("A","B")), C = matrix(1:6,ncol = 3))
list3 = list(A = 1:4, B = as.factor(c("A","B","C")),
             C = matrix(c(1:5,7),ncol = 3))
list4 = list(A = 1:4, B = as.factor(c("A","B","C")), C = matrix(1:6,ncol = 2))
list5 = list(A = 1:4, B = as.factor(c("A","B")), C = matrix(1:6,ncol = 2))
list6 = list(A = 1:4, B = as.factor(c("A","B")), C = data1)
list7 = list(A = 1:4, B = as.factor(c("A","B")), C = data2)

compare_lists(list1,list2)
compare_lists(list1,list3)
compare_lists(list1,list4)
compare_lists(list1,list5)
compare_lists(list6,list7)
}
