% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_pos.r
\name{count_pos}
\alias{count_pos}
\title{count_pos()}
\usage{
count_pos(n, n_grp)
}
\arguments{
\item{n}{An integer. The cardinal of a set}

\item{n_grp}{An integer. The number of desired subsets to compose the initial set}
}
\value{
A list of possible cardinals to compose the initial sets with n_groups
}
\description{
A function that counts the cardinals of possible subgroups from a set
}
\examples{
a1 = count_pos(5,2); a1

}
