% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.R
\name{cout}
\alias{cout}
\title{Fonction COUT}
\usage{
cout(dat, choix_dist = "E")
}
\arguments{
\item{dat}{BDD de la forme [base|Y1|Y2|???], ??? ?tant des covariables
binaires ou quantis (apr?s transformation par fct prepar-cov()
par exemple)}

\item{choix_dist}{(FACULTATIF) Distance euclidienne "E" par d?faut, sinon "H" pour
distance de Hamming}
}
\description{
Cette fonction est la partie I/II de la méthode FOT
Elle estime les co?ts de passage entre les modalit?s Y1 et Y2
}
