% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_part.r
\name{family_part}
\alias{family_part}
\title{family_part()}
\usage{
family_part(n, ordinal = FALSE)
}
\arguments{
\item{n}{An integer. The cardinal of the set}

\item{ordinal.}{A boolean. If TRUE the power set is only composed of subsets of consecutive elements, FALSE (by default) otherwise}
}
\value{
A list of (2^n -1) subsets (The empty set is exclude)
}
\description{
Function that gives the power set of a set
}
\examples{
fam  =  family_part(4)
fam2 =  family_part(4,ordinal = TRUE)
}
