% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_coord.r
\name{find_coord}
\alias{find_coord}
\title{find_coord()}
\usage{
find_coord(all_ind, all_list, all_coord)
}
\arguments{
\item{all_ind}{A vector of integers from the count_pos() function}

\item{all_list}{A list of sublists of subsets of a power set (from family_part() by example) composed of 1,2,...,n elements}

\item{all_coord}{A matrix of integers which lines gives the coordinates of a list from family_part()}
}
\value{
indices of a list from family_part()
}
\description{
A function that returns the indices of a list from family_part() to compose the partition of a set
}
\examples{
fam      =  family_part(4)
size_grp = sapply(fam,length)
a1       = count_pos(4,3)
grp_list = list()
ind_list = list()
ind      = 1:3
for (k in 1:3){
  grp_list[[k]] = fam[size_grp == ind[k]]
  ind_list[[k]] = (1:length(fam))[size_grp == ind[k]]
}
list_coord = find_coord(unlist(a1[[1]]),grp_list,ind_list)

}
