% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{gendata}
\alias{gendata}
\title{SIMULATION DE DONNEES POUR APPROCHE DETERMINISTE
 
  -> Y1 et Y2 sont enti?rement d?termin?es par les valeurs de X1 et X2
  -> Cette fonction g?n?re une BDD de la forme [Base|Y1|Y2|X1|X2] si X1 et X2
     sont cat?gorielles ? 2 modalit?s en m?me temps.
  -> Cette fonction g?n?re une BDD de la forme [Base|X1|X2] si X1 et X2
     sont cat?gorielles avec X1 ou X2 comptent plus de 2 modalit?s}
\usage{
gendata(n1, k, rho, p1c, p2c)
}
\arguments{
\item{n1}{Nombre de sujets dans la base 1}

\item{k}{Coefficient de proportionnalit? entre le nbr de sujets dans la Base 1 et le nombre de sujets dans la Base 2}

\item{rho}{Coefficient de corr?lation entre les covariables X1 et X2}

\item{p1c}{Distribution de la covariable X1 cat?gorielle (format: c(,))c}

\item{p2c}{Distribution de la covariable X2 cat?gorielle (format: c(,))}
}
\description{
SIMULATION DE DONNEES POUR APPROCHE DETERMINISTE
 
  -> Y1 et Y2 sont enti?rement d?termin?es par les valeurs de X1 et X2
  -> Cette fonction g?n?re une BDD de la forme [Base|Y1|Y2|X1|X2] si X1 et X2
     sont cat?gorielles ? 2 modalit?s en m?me temps.
  -> Cette fonction g?n?re une BDD de la forme [Base|X1|X2] si X1 et X2
     sont cat?gorielles avec X1 ou X2 comptent plus de 2 modalit?s
}
\examples{

essai1 = gendata(20,1,0.3,p1c = c(0.2,0.4,0.1,0.3),p2c = c(0.6,0.4))
essai2 = gendata(20,1,0.3,p1c = c(0.6,0.4),p2c = c(0.7,0.3))
}
