% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imput_cov.r
\name{imput_cov}
\alias{imput_cov}
\title{imput_cov()}
\usage{
imput_cov(
  dat1,
  indcol = 1:ncol(dat1),
  R_mice = 5,
  meth = rep("pmm", ncol(dat1)),
  missMDA = FALSE,
  NB_COMP = 3,
  seed_choice = sample(1:1e+06, 1)
)
}
\arguments{
\item{dat1}{A data.frame. containing the variables to be imputed and the one participating in the imputations}

\item{indcol}{A vector of integers. The corresponding column numbers corresponding to the variables to be imputed and those which participate in the imputations.}

\item{R_mice}{An integer. The number of imputed database to generate with MICE method (5 by default).}

\item{meth}{A vector of characters which specify the imputation method to be used for each column in dat1.
"pmm" for continuous covariates only, "logreg" for binary covariates, "polr" for ordinal covariates, "polyreg" for categorical covariates (no order),(cf \code{\link[mice]{mice}} for more details).}

\item{missMDA}{A boolean. If \code{TRUE}, missing values are imputed using the factoral analysis for mixed data (\code{\link[missMDA]{missMDA-package}}) from the missMDA package (Simple imputation).}

\item{NB_COMP}{An integer corresponding to the number of components used to predict the missing entries (3 by default) when the \code{missMDA} option is TRUE.}

\item{seed_choice}{An integer used as argument by the set.seed() for offsetting the random number generator (Random integer by default)}
}
\value{
A list of 3 or 4 objects (depending on the missMDA option. 3 objects if \code{missMDA} = TRUE, otherwise 4 objects are returned:
\item{RAW}{A data.frame corresponding to the raw database}
\item{IMPUTE}{A character corresponding to the type of imputation selected}
\item{DATA_IMPUTE}{A data.frame corresponding to the completed (consensus if multiple imputations) database}
\item{MICE_IMPS}{Only if missMDA = FALSE. A list object containing the R imputed databases generated by MICE}
}
\description{
A function that performs imputations on incomplete covariates, whatever their types, using functions from the \pkg{MICE} package (Van Buuren's Multiple Imputation) or functions from the \pkg{missMDA} package (Simple Imputation wiht Multivariate data analysis)
}
\details{
When multiple imputations are required (for MICE only), each missing information is imputed by a consensus value.
The average of the candidate values will be retained for numerical variables, while the most frequent class will be retained for categorical variables (ordinal or not).
The \code{MICE_IMPS} output stored the imputed databases to allow users to buid other consensus values by themselves, and to assess the variabilities related to the imputed values, if necessary.
}
\examples{
# By example,we decide to impute all incomplete covariates in the simu_data table.
data(simu_data)

# Here We keep the complete \code{Gender} covariate in the imputation model.
# Using MICE (REP = 3):
imput_mice = imput_cov(simu_data,indcol = 4:8,R_mice = 3,
                       meth = c("logreg","polyreg","polr","logreg","pmm"))
summary(imput_mice)


# Using FAMD (NB_COMP = 3):
imput_famd = imput_cov(simu_data,indcol = 4:8,
                       meth = c("logreg","polyreg","polr","logreg","pmm"),
                       missMDA = TRUE)
summary(imput_famd)

}
\references{
Audigier, V., Husson, F. & Josse, J. (2013). A principal components method to impute mixed data. Advances in Data Analysis and Classification, 10(1), 5-26.

van Buuren, S., Groothuis-Oudshoorn, K. (2011). mice: Multivariate Imputation by Chained Equations in R. Journal of Statistical Software, 45(3), 1–67.
}
\author{
Gregory Guernec
\email{gregory.guernec@inserm.fr}
}
