% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_from_group_optimal.r
\name{individual_from_group_optimal}
\alias{individual_from_group_optimal}
\title{individual_from_group_optimal(inst, jointprobaA, jointprobaB, percent_closest=1.0)}
\usage{
individual_from_group_optimal(inst, jointprobaA, jointprobaB,
  percent_closest = 1)
}
\arguments{
\item{inst}{}

\item{jointprobaA}{}

\item{jointprobaB}{}

\item{percent_closest}{}
}
\value{
list
}
\description{
individual_from_group_optimal(inst, jointprobaA, jointprobaB, percent_closest=1.0)
}
