% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instance.r
\name{Instance}
\alias{Instance}
\title{Instance(data_file,norme)}
\usage{
Instance(data_file, norme)
}
\arguments{
\item{data_file}{todo list}

\item{norme}{todo list}
}
\value{
todo list
}
\description{
Instance(data_file,norme)
}
