% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OT.r
\name{OT}
\alias{OT}
\title{OT(datab)}
\usage{
OT(datab, percent_c = 1, maxrelax = 0, norm = 1, indiv_method,
  full_disp = FALSE, solver_disp = FALSE)
}
\arguments{
\item{datab}{todo list}

\item{percent_c}{todo list}

\item{maxrelax}{todo list}

\item{norm}{todo list}

\item{indiv_method}{todo list}

\item{full_disp}{todo list}

\item{solver_disp}{todo list}
}
\value{
list with time and solution
}
\description{
OT(datab)
}
