% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_pred.r
\name{select_pred}
\alias{select_pred}
\title{select_pred()}
\usage{
select_pred(
  databa,
  Y,
  ID = 1,
  quanti = NULL,
  nominal = NULL,
  ordinal = NULL,
  logic = NULL,
  convert_num = NULL,
  convert_clss = NULL,
  thresh_cat = 0.3,
  thresh_num = 0.7,
  thresh_Y = 0.2,
  RF = TRUE,
  RF_ntree = 500,
  RF_condi = FALSE,
  RF_condi_thr = 0.2,
  RF_SEED = sample(1:1e+06, 1)
)
}
\arguments{
\item{databa}{A data.frame with a column of identifiers, an outcome, and a set of predictors. The number of columns can exceed the number of rows.}

\item{Y}{The label of the outcome with quotes}

\item{ID}{The index of column of the row identifiers}

\item{quanti}{A vector of integers which corresponds to the indexes of columns of all the numeric predictors}

\item{nominal}{A vector of integers which corresponds to the indexes of columns of all the categorical nominal predictors}

\item{ordinal}{A vector of integers which corresponds to the indexes of columns of all the categorical ordinal predictors}

\item{logic}{A vector of integers indicating the indexes of logical predictors. No index remained by default}

\item{convert_num}{A vector of integers indicating the indexes of quantitative variables to convert in ordered factors. No index remained by default. Each index selected must has been defined as quantitative in the quanti argument.}

\item{convert_clss}{A vector of integers indicating the number of classes chosen for each transformation of quantitative variable in ordered factor. The length of this vector can not exceed the length of the convert_num argument. If length(convert_num) > 1 and length(convert_clss) = 1.
All quantitative predictors selected for discretisation will have the same number of classes.}

\item{thresh_cat}{A threshold associated to the Cramer's V coefficient (= 0.30 by default)}

\item{thresh_num}{A threshold associated to the Spearman's coefficient of correlation (= 0.70 by default)}

\item{thresh_Y}{A threshold indicating the minimal percent of importance required for a variable to be considered as a reliable predictor}

\item{RF}{A boolean that equals TRUE (default) if a random forest procedure must be used in the selecting of the best subset of predictors for the outcome}

\item{RF_ntree}{Number of bootsrap samples required from the row datasource during the random forest procedure}

\item{RF_condi}{A boolean specifying if the conditional importance measures must be assessed from the random forest procedure (\code{TRUE}) rather than the standard variable importance  measures (\code{FALSE} by default)}

\item{RF_condi_thr}{A threshold linked to (1 - pvalue) of an association test, kwnowing that a threshold value of zero wil include all covariates.}

\item{RF_SEED}{An integer used as argument by the set.seed() for offsetting the random number generator (Random integer by default)}
}
\value{
A list of 13 (If \code{RF} = TRUE) or 10 objects (Only the 1st ten objects if \code{RF} = FALSE) is returned:
\item{seed}{The random number generator fixed or selected}
\item{thresh}{A summarize of the different thresholds fixed for the study}
\item{convert_num}{The labels of the continuous predictors transformed in categorical form}
\item{DB_USED}{The final database used after eventual transformations of predictors}
\item{vcrm_Y_cat}{Table of pairwise associations between the outcome and the categorical predictors (Cramer's V)}
\item{cor_Y_num}{Table of pairwise associations between the outcome and the continuous predictors (Rank correlation)}
\item{v_crm_X_cat}{Table of pairwise associations between the categorical predictors (Cramer's V)}
\item{cor_X_num}{Table of pairwise associations between the continuous predictors (Cramer's V)}
\item{FG_test}{Results of the Farrar and Glauber tests, with and without approximation form}
\item{colinear_PB}{Table of predictors with problem of collinearity}
\item{drop_var}{Labels of predictors to drop after RF process (optional output: Only if \code{RF}=TRUE)}
\item{RF_PRED_Y}{Table of variable importance measurements, conditional or not, according to the argument \code{condi_RF} (optional output: Only if \code{RF}=TRUE)}
\item{best_pred}{Labels of the best predictors selected (optional output: Only if \code{RF}=TRUE)}
}
\description{
Selection of a subset of non collinear predictors having relevant relationships with a given target outcome.
}
\details{
The \code{select_pred} function provides several tools to identify, on the one hand, the relationships between predictors, by detecting especially potential problems of collinearity, and, on the other hand, proposes a parcimonious subset of relevant predictors using appropriate random forest procedures.
Even if, in many contexts, this function can be seen as a preliminary step of regression, this function is particularly adapted to the context of data integration of two or more datasources by pre-selecting relevant subsets of shared predictors before fusion.


A. STRUCTURE OF DATABASE REQUIRED

The input expected database is a data.frame that especially requires a specific column of rows identifiers and a target variable (or outcome) having a finite number of values or classes (Therefore,in preference, of ordinal, nominal or discrete type). If the chosen outcome is in numeric form, it will be automatically converted in ordinal type.
The number of predictors is not a constraint (even if, with less than 3 variables a process of variables selection has no real sense...), can exceed the number of rows (no problem of high dimensionality here).
The predictors can be continuous (quantitative), nominal or ordinal with or without missing values. Nevertheless, this function provides optimal performances without or with only a limited number of continuous variables and with complete information. For this reason, this function is particularly adapted to
data integration using the \code{OT_joint} function available in this package, which imposes the same conditions.
In presence of numeric variables, it is therefore suggested to discretize them beforehand or to use the process of discretization directly integrated in the function. For doing this, \code{convert_num} and \code{convert_class} are the only two arguments that have to be strictly completed. With the \code{convert_num} argument, users choose the
continuous variables to convert and the \code{convert_clss} argument specify the related number of classes chosen for each discretization.
So, these 2 arguments must be 2 vectors of indexes of same length. The only exception allowed, when \code{convert_clss} equals a scalar S, supposed that the continuous predictors selected for conversion  will be disretized with a same number of classes S.
By example, if \code{convert_class} = 4, all the continuous variables selected with the \code{convert_num} argument will be discretized by quartiles. Finally, missing informations from a predictor to convert, are not discretized and stay unchanged during the transformation.


B. PAIRWISE ASSOCIATIONS BETWEEN PREDICTORS

In a first step of process, \code{select_pred} calculates standard pairwise associations between predictors according to their types.
\enumerate{
\item{Between categorical predictors (ordinal, nominal and logical):
Cramer's V (and Bias-corrected Cramer's V, see Bergsma 2013, for more details) are calculated between categorical predictors and the \code{thres_cat} argument fixed the associated threshold beyond which two predictors can be considered as redundant.
A similar process is done between the target variable and the subset of categorical variables which provides in output, a first table ranking the top scoring predictor variates which summarizes their abilities to predict the target.}
\item{Between continuous predictors:
If the\code{ordinal} and \code{logic} arguments differ from NULL, all the corresponding predictors are beforehand converted in discrete form.
For numeric (quantitative), logical and ordinal predictors, pairwise correlations between ranks (Spearman) are calculated and the \code{thresh_num} arguments fixed the associated threshold beyond which two predictors can be considered as redundant.
A similar process is done between the outcome and the subset of discrete variables which provides in output, a table ranking the top scoring predictor variates which summarizes their abilities to predict the target.
In addition, the result of a Farrar and Glauber test is provided (see the corresponding reference for more details). This test is based on the determinant of the correlation matrix of covariates and the related null hypothesis of the test corresponds to an absence of collinearity between them.
In presence of a large number of numeric covariates and/or ordered factors, the approximate Farrar-Glauber test, based on the normal approximation of the null distribution is more adapted (Johnson, 1994) and its result is also provided in output.
These 2 tests are highly sensitive and, by consequence, it suggested to consider these results as simple indicators of collinearity between predictors rather than an essential condition of acceptability.}
}
If the initial number of predictors is not too important, these informations can be sufficient to users for the visualization of the potential problems of collinearity and for the selection of a subset of predictors (\code{RF} = FALSE).
It is nevertheless often necessary to complete this visualization by an automatical process of selection like the Random Forest approach (see Breiman 2001, for more details about RF).


C. RANDOM FOREST PROCEDURE

As a final step of the process, a random forest approach (RF) is here prefered (to regressions) for its great flexibility: A non parametric recursive partitioning method applyable even when the number of variables exceeds the number of rows, whatever the types of covariates considered, able to deal with non linearity,
correlated predictors, ordered or not ordered categorical outcomes. For this, \code{select_pred} integrates in its algorithm the \code{\link[party]{cforest}} and \code{\link[party]{varimp}} of the package \pkg{party} (Hothorn, 2006) and so gives access
to its main arguments.

A RF approach generally provides 2 types of measures for estimating the mean variable importance of each covariate in the prediction of an outcome: The Gini importance and the permutation importance.These measurements must be used with caution, by taking into account the following constraints:
\enumerate{
\item{The Gini importance criterion can produce bias in favor of continuous variables and variables with many categories. To avoid this problem, only the permutation criterion is available in the function.}
\item{The permutation importance criterion can overestimate the importance of highly correlated predictors.}
}
The function \code{select_pred} can so proceeds with 3 different scenarios according to the types of predictors:
\enumerate{
\item{The 1st scenario consists in boiling down to a set of categorical variables (ordered or not) by discretizing all the continuous predictors beforehand, using the internal \code{convert_num} argument or another one, and then works with the conditional importance measurements (\code{RF_condi} = TRUE) which give unbiased estimations.
In the spirit of a partial correlation, conditional importance measure of a variable X, only use a more or less important part of predictors (called Z) correlated to X for the computation of the estimation. The argument \code{RF_condi_thr} that corresponds exactly to the \code{threshold} argument of the function \code{\link[party]{varimp}},
fixes this ratio according to an association test between X and Z (see Strobl 2007 for more precision). It corresponds to 1 - pvalue of the test, kwnowing that a threshold value of zero will include all predictors in the computation.
Moreover, taking into account only subsets of predictors in the computation of the variable importance measures could lead to a relevant saving of execution time.
Nevertheless, this method does not take into account incomplete predictors an so have to be imputed beforehand if necessary (using the \code{imput_cov} function available in this package or another one).}
\item{The 2nd possibility, always in presence of mixed types predictors, consists in the execution of 2 successive RF procedures. The 1st one is used to select an unique candidate in each susbset of correlated predictors (detecting in the first section), and the 2nd one extracts the permutation importance
measurements from a second RF applies on the remaining subset of uncorrelated predictors (\code{RF_condi} = FALSE, default option). This 2nd possibility has the advantage to work in presence of incomplete predictors.}
\item{The 3rd scenario consists in running a first time the function without RF process (\code{RF} = FALSE), and according to the presence of highly correlated predictors or not, makes a decision about the type of criterion used as a 2nd execution of the function.}
}
The 3 scenarios finally provide a list of uncorrelated predictors of the outcome sorted in unbiased importance order and the argument \code{thresh_Y} corresponds to the minimal percent of importance required for a variable to be considered as a reliable predictor of the outcome.
Finally, because all random forest results are subject to random variation, users can check whether the same importance ranking is achieved with a different random seed (\code{RF_SEED} argument) or otherwise by increasing the number of trees (\code{RF_ntree}).
}
\examples{
# Using the simu_data table, firstly, we separate the 2 bases A and B
# and remove the unknown outcome in each DB
data(simu_data)
simu_A = simu_data[simu_data$DB == "A",-3]    # Base A
simu_B = simu_data[simu_data$DB == "B",-2]    # Base B

# In the 1st DB(A), the target variable is Yb1.
# The indexes of the column covariates goes from 3 to 7.
# By keeping the other default parameters:

\dontrun{
### Scenario 1: Discretization of age (7) in 3 classes (so no quantitative predictors),
# RF process chosen, and conditional importance measures assessed, all other arguments
# take their default values:
sel_A2 = select_pred(simu_A,"Yb1", quanti = 7, nominal = c(3,4,6), ordinal = c(2,5),
                     convert_num = 7,convert_clss = 3,
                     RF = TRUE, RF_condi = TRUE)


### Scenario 2: Raw predictors, RF process chosen, and standard permutation
# importance measures assessed (Default values for other arguments)
sel_A3 = select_pred(simu_A,"Yb1", quanti = 7, nominal = c(3,4,6), ordinal = c(2,5),
                     RF = TRUE)
}

### Scenario 3: Discretization of age (7) in 3 classes and no random forest process:
sel_A1 = select_pred(simu_A,"Yb1", quanti = 7, nominal = c(3,4,6), ordinal = c(2,5),
                     convert_num = 7,convert_clss = 3,RF = FALSE)


}
\references{
# About the Cramer's V measures:
Bergsma W (2013). A bias-correction for Cramer's V and Tschuprow's T. Journal of the Korean Statistical Society, 42, 323–328.

# About the Farrar and Glauber test:
Farrar D, and Glauber R (1968). Multicolinearity in regression analysis. Review of Economics and Statistics, 49, 92–107.

# About RF:
Breiman L (2001). Random Forests. Machine Learning, 45(1), 5–32.

About the estimation of conditional variable importances in RF:
Strobl C, Boulesteix A-L, Kneib T, Augustin T, Zeileis A (2008). Conditional Variable Importance for Random Forests. BMC Bioinformatics, 9, 307.
}
\author{
Gregory Guernec
\email{gregory.guernec@inserm.fr}
}
