% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.R
\name{simpond}
\alias{simpond}
\title{simpond(x)
Cette fonction propose des arrondis aux effectifs de la matrice de simplexe
lorsque cette dernière n'est pas uniquement composée d'entiers
En cas d'ex-aecquos, le premier pr?sent? est le premier servi}
\usage{
simpond(x)
}
\arguments{
\item{x}{un vecteur}
}
\value{
un vecteur aussi
}
\description{
simpond(x)
Cette fonction propose des arrondis aux effectifs de la matrice de simplexe
lorsque cette dernière n'est pas uniquement composée d'entiers
En cas d'ex-aecquos, le premier pr?sent? est le premier servi
}
\examples{
x = c(0.5,0,0.5,0.5,3.5); x; sum(x)
simplex_new = simpond(x); simplex_new; sum(simplex_new)
simplex_new = simpond(x); simplex_new; sum(simplex_new)
}
