% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simul_glob}
\alias{simul_glob}
\title{IV.  FONCTIONS D'ESTIMATION DES ERREURS
Cette fonction retourne sur une ligne, les erreurs moyennes de pr?diction
des m?thodes FOT et MICE appliqu?es sur donn?es simul?es par approche
d?terministe ou non-d?terministe}
\usage{
simul_glob(n_1 = NULL, coeff = NULL, p1cc = NULL, p2cc = NULL,
  qual = NULL, quan = NULL, choose_dist = "E", rho_cov = NULL,
  rep_MICE = 5, cov_MICE = "pmm", calcu = "FREQ", typ_simu = NULL,
  r2t = NULL, py1cc = NULL, py2cc = NULL, vlY1 = NULL,
  vlY2 = NULL)
}
\arguments{
\item{n_1}{: Taille de la base 1}

\item{coeff}{: Coefficient de proportionnalit? taille base 2}

\item{p1cc}{: Proportion des cat?gories pour la covariable 1}

\item{p2cc}{: Proportion des cat?gories pour la covariable 1}

\item{qual}{: Les n? de colonnes des variables qualitatives}

\item{quan}{: Les n? de colonnes des variables quantitatives}

\item{rep_MICE}{: Nbre de BDD ? imputer pour MICE}

\item{cov_MICE}{: Sp?cifier si Y1, Y2, X1, X2 et X3 doivent ?tre consid?r?es dans
MICE au format num?rique ("pmm") ou facteur ("polyreg")}

\item{calcu}{: "FREQ" ou "MOY" selon que les variables imput?es soient quantis
ou qualis}

\item{typ_simu}{: D?terministe "D" ou Non D?terministe "ND"}

\item{r2t}{: R^2 de la r?gression lin?aire Ycontinue = f(X1,X2,X3)}

\item{py1cc}{: Proportion de patients dans chaque classe de Y en base 1}

\item{py2cc}{: Proportion de patients dans chaque classe de Y en base 2}

\item{vlY1}{: Les labels de chaque classe pour Y cat?gorielle en base 1}

\item{vlY2}{: Les labels de chaque classe pour Y cat?gorielle en base 2}

\item{M}{: Nombre de r?p?titions}

\item{choose_dist:}{Choix de la distance pour la m?thode FOT ("E" ou "H")}

\item{Les}{coefficients suivants ne sont utilis?s que dans le cas non-d?terministe:}

\item{bdd}{: Jeu de donn?es non simul?}

\item{seedsim}{: G?n?rateur al?atoire (tir? au sort par d?faut)}
}
\value{
missing documentation
}
\description{
IV.  FONCTIONS D'ESTIMATION DES ERREURS
Cette fonction retourne sur une ligne, les erreurs moyennes de pr?diction
des m?thodes FOT et MICE appliqu?es sur donn?es simul?es par approche
d?terministe ou non-d?terministe
}
\examples{
n.sim = 1
n_1         = 1000
coeff       = 1
p1cc        = c(0.5,0.5)
p2cc        = c(0.3,0.3,0.4)
qual        = 4:5
quan        = 6
choose_dist = "E"
rho_cov     = 0.2
rep_MICE    = 5
cov_MICE    = c("polyreg","polyreg","polyreg","polyreg","pmm")
calcu       = "FREQ"
typ_simu    = "ND"
r2t         = 0.5
py1cc       = c(0.25,0.25,0.25,0.25)
py2cc       = c(0.333,0.334,0.333)
vlY1        = c(1,2,3,4)
vlY2        = c(1,2,3)
#simul_glob (n_1, coeff, p1cc, p2cc, qual , quan ,
#            choose_dist="E", rho_cov, rep_MICE=5, cov_MICE = "pmm", calcu="FREQ",
#            typ_simu,r2t , py1cc, py2cc,
#           vlY1, vlY2)
}
