% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{transfo}
\alias{transfo}
\title{fonction TRANSFO
  Cette fonction transforme une variables qualitative (format facteur) ? k
  classes en (k-1) variables binaires}
\usage{
transfo(x, nom = NULL)
}
\arguments{
\item{x}{Variable facteur}

\item{nom}{Possibilite de donner un nom a votre variable pour mieux reperer les
binaires correspondantes en sortie
La modalite de reference est la 1ere (cf levels(x) au prealable)}
}
\description{
fonction TRANSFO
  Cette fonction transforme une variables qualitative (format facteur) ? k
  classes en (k-1) variables binaires
}
