% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfo_dist.r
\name{transfo_dist}
\alias{transfo_dist}
\title{transfo_dist()}
\usage{
transfo_dist(DB, index_DB_Y_Z = 1:3, quanti = NULL, nominal = NULL,
  ordinal = NULL, logic = NULL, prep_choice = "E", info = 80)
}
\arguments{
\item{DB}{A data.frame composed of exactly 2 superimposed databases with a column of database identification, 2 columns corresponding to a same information
differently encoded in the 2 databases and covariates. The order of the variables have no importance.}

\item{index_DB_Y_Z}{A vector of exactly 3 integers. The 1st integer must correspond to the index of the database identifiaction column. The 2nd integer corresponds
to the index of the target variable in the 1st database while the 3rd integer corresponds to the index of column related to the target varaible in the 2nd database.}

\item{quanti}{A vector of integers that corresponds to the indexes of columns of all the quantitative variables (DB identification and target variables included)}

\item{nominal}{A vector of integers that corresponds to the indexes of columns of all the nominal (not ordered) variables (DB identification and target variables included)}

\item{ordinal}{A vector of integers that corresponds to the indexes of columns of all the ordinal variables (DB identification and target variables included)}

\item{logic}{A vector of integers that corresponds to the indexes of columns of all the boolean variables.}

\item{prep_choice}{A character (with quotes) corresponding to the distance function chosen between: The euclidean distance ("E", by default), The Manhattan distance ("M"),
the Gower distance ("G"), the Hamming (also called binary) distance and the Euclidean or Manhattan distance, calculated from principal components of a factor analysis of mixed data ("FAMD").}

\item{info}{A percent value (between 0 and 100) that corresponds to the part of variability taken into account by the principal components of the FAMD when this option is required.}
}
\value{
A data.frame which covariates have been transformed according to the distance function chosen. The columns of the data.frame could have been reordered so that the identifier, Y and Z correspond to the 1st three columns respectively.
}
\description{
A function that prepares the database according to the distance function chosen to evaluate the proximities between rows of databases in the OT algorithm
}
\details{
The database must have at least 4 columns (stored in a nonspecific order) : An identifier column (of 2 classes) to distinguish the 2 superimposed databases,
a (nominal or ordinal) variable corresponding to the information to merge with its specific encoding in the 1st database (called Y by example),
the corresponding target variable in the 2nd database (called Z by example), and at least a covariate common to both databases (that is to say with same encoding in the 2 bases).

In this context, the information related to Y in the 2nd database is missing as the information related to Z in the 1st database.
We remind users that the purpose of this package is to predict the missing informations in Y and Z.

All column indexes (including those related to identifier and target variables Y and Z) of the used database must be declare once, among the \code{quanti}, \code{nominal}, \code{ordinal}, or \code{logic} options.

TRANSFORMATIONS ON THE DATABASE ACCORDING TO THE CHOICE OF THE DISTANCE
-----------------------------------------------------------------------

These necessary transformations are related to the type of each of the covariates.
Obviously it depends on the choice of the distance function chooses by user in the \code{prep_choice} option.

1. For the Euclidean (E) and Manhattan (M) distances:
The numeric variables are unchanged and the potential transformations only concerned all other types of covariates (if there is any).
The related recoding to a boolean variable is 1 for \code{TRUE} and 0 for \code{FALSE}.
The recoding for a nominal variable of k classes corresponds to its related disjunctive table (i.e (k-1) binary variables))
The ordinal variables are all converted to numeric variables (Please take care that the order of the classes of each of these variables is well specified at the beginning).

2. For the Hamming (H) distance:
All the numeric variables must be categorize beforehand as described in the 4th example.
The related recoding to a boolean variable is 1 for \code{TRUE} and 0 for \code{FALSE}
The recoding for nominal or ordinal variable of k classes corresponds to its related disjunctive table (i.e (k-1) binary variables))

3. For the Gower (G) distance:
All covariates remain unchanged

4. Using the principal components from a factor analysis for mixed data (FAMD):
A factor analysis for mixed data is done on the covariates of the database and a specific number of the related principal components is remained (depending on the variability part explained by the covariates that the user wishes to keep by varying the \code{info} option).
After this step, all new covariates will be obviously in numeric forms.
}
\examples{

### Using the simu_data example, suppose that you want to prepare
### your data before applying the OT algorithm with:

data(simu_data)

# 1. the Euclidean distance (same output with Manhattan distance),
try1 = transfo_dist(simu_data,quanti = c(3,8), nominal = c(1,4:5,7),
                    ordinal = c(2,6), logic = NULL, prep_choice = "E")
# Here Yb2 was stored in numeric: It has been automatically converted in factor

# You can also convert beforehand Yb2 in ordered factor by example:
sim_data     = simu_data
sim_data$Yb2 = as.ordered(sim_data$Yb2)
try1 = transfo_dist(sim_data,quanti = 8, nominal = c(1,4:5,7),
                    ordinal = c(2,3,6), logic = NULL, prep_choice = "E")

# 2. The Euclidean distance on principal components generated
#    by a factor analysis for mixed data (FAMD):
try2 = transfo_dist(simu_data,quanti = c(3,8), nominal = c(1,4:5,7),
                    ordinal = c(2,6), logic = NULL, prep_choice = "FAMD")

# Please notice that this method works only with rows that have complete
# information on covariates.

# 3. The Gower distance for mixed data:
try3 = transfo_dist(simu_data,quanti = c(3,8), nominal = c(1,4:5,7),
                    ordinal = c(2,6), logic = NULL, prep_choice = "G")

# 4. The Hamming distance:
# Here the quanti option could only contain indexes related to targets.
# Indexes columns related to potential binary covariates or covariates with
# finite number of integer values must be include in the ordinal option.
# So in simu_data, the Age covariate can not be directly use without prior
# categorization.

simu_dat = simu_data[,-8]    # categorization
simu_dat$AgeC = cut(simu_data$Age,breaks = c(34,50,54,66))
try4 = transfo_dist(simu_dat,quanti = 3, nominal = c(1,4:5,7),ordinal = c(2,6,8),
prep_choice = "H")

# Other situation with a new numeric covariate with a finite number of values
simu_dat = simu_data[,-8]
simu_dat$new = sample(1:3,nrow(simu_data),replace = TRUE)
try5 = transfo_dist(simu_dat,quanti = 3, nominal = c(1,4:5,7),ordinal = c(2,6,8),
prep_choice = "H")


### This function works whatever the order of your columns in your database:
# Suppose that we re-order columns in simu_data:
simu_data2 = simu_data[,c(2,4:7,3,8,1)]

# By changing the corresponding indexes in the index_DB_Y_Z option,
# we observe the desired output:
try5 = transfo_dist(simu_data2,index_DB_Y_Z = c(8,1,6),quanti = 6:7, nominal = c(2:3,5,8),
                     ordinal = c(1,4), logic = NULL, prep_choice = "E")

}
\references{
# For Factor Analysis with mixed data:
Pages J. (2004). Analyse factorielle de donnees mixtes. Revue Statistique Appliquee. LII (4). pp. 93-111.

# About the Gower distance:
Gower, J. C. (1971), “A general coefficient of similarity and some of its properties”. Biometrics, 27, 623--637.

# About the other distance measurements:
Anderberg, M.R. (1973), Cluster analysis for applications, 359 pp., Academic Press, New York, NY, USA.

Borg, I. and Groenen, P. (1997) Modern Multidimensional Scaling. Theory and Applications. Springer.
}
\author{
Gregory Guernec
\email{gregory.guernec@inserm.fr}
}
