% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfo_quali.r
\name{transfo_quali}
\alias{transfo_quali}
\title{transfo_quali()}
\usage{
transfo_quali(x, labx = NULL)
}
\arguments{
\item{x}{a factor}

\item{labx}{a new label for the generated binary variables (By default the name of the factor is conserved)}
}
\value{
A matrix of (n-1) binary variables
}
\description{
A function that transforms a factor of n(>1) levels in (n-1) binary variables.
}
\examples{

treat     = as.factor(c(rep("A",10),rep("B",15),rep("C",12)))
treat_bin = transfo_quali(treat,"trt")


}
\author{
Gregory Guernec

\email{otrecod.pkg@gmail.com}
}
