% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfo_target.r
\name{transfo_target}
\alias{transfo_target}
\title{transfo_target()}
\usage{
transfo_target(z, levels_order = NULL)
}
\arguments{
\item{z}{A vector}

\item{levels_order}{A vector of levels. When the target is ordinal, the levels can be sorted by ascending order.
By default, the initial order is retained.}
}
\value{
A list containing two elements:
        NEW An object of class factor of the same length as z
        LEVELS_NEW The levels (ordered or not) retained for z
}
\description{
A function which prepare the encoding of the target variable before OT algorithm
}
\examples{
y      = rnorm(100,30,10)
aa     = transfo_target(y)

newlev = unique(as.integer(y))
bb     = transfo_target(y,levels_order = newlev)
newlev2 = newlev[-1]
cc      = transfo_target(y,levels_order = newlev2)

outco   = c(rep("A",25),rep("B",50),rep("C",25))
dd      = transfo_target(outco,levels_order = c("B","C","A"))
ee      = transfo_target(outco,levels_order = c("E","C","A","F"))
ff      = transfo_target(outco)

outco2  = c(rep("A",25),NA,rep("B",50),rep("C",25),NA,NA)
gg      = transfo_target(outco2)
hh      = transfo_target(outco2,levels_order = c("B","C","A"))
}
