% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_group.r
\name{try_group}
\alias{try_group}
\title{try_group()}
\usage{
try_group(Y1, Y2, ordin = FALSE)
}
\arguments{
\item{Y1}{outcomes first database}

\item{Y2}{outcomes second database}

\item{ordin}{family_part option}
}
\value{
list
}
\description{
try_group()
}
\examples{
Z1 = as.factor(sample(1:3,50,replace = TRUE)); length(Z1)
Z3 = as.factor(sample(1:2,50,replace = TRUE)); length(Z3)
Z2 = as.factor(sample(c("A","B","C","D"),50, replace = TRUE)); length(Z2)
Z4 = as.factor(sample(c("A","B","C","D","E"),50, replace = TRUE)); length(Z4)

a1 = try_group(Z1,Z2,ordin = FALSE)
summary(a1); a1[[1]]
aa = try_group(Z3,Z2,ordin = FALSE)
bb = try_group(Z4,Z2,ordin = TRUE)
cc = try_group(Z4,Z1,ordin = TRUE)

}
