% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif_OT.r
\name{verif_OT}
\alias{verif_OT}
\title{verif_OT()}
\usage{
verif_OT(
  ot_out,
  group.clss = FALSE,
  ordinal = TRUE,
  stab.prob = FALSE,
  min.neigb = 1,
  R = 10,
  seed.stab = sample(1:1e+06, 1)
)
}
\arguments{
\item{ot_out}{An output object of the function \code{OT_outcome} or \code{OT_joint}}

\item{group.clss}{A boolean indicating if proximity between outcomes is assessed by grouping levels of factor (FALSE by default)}

\item{ordinal}{A boolean that indicates if Y and Z are ordinal (\code{TRUE} by default) or not. This argument is only used in the context of grouping of level (\code{group.clss}=TRUE).}

\item{stab.prob}{A boolean indicating if the results related to the stability of the predictions are requested.}

\item{min.neigb}{A value indicating the minimal number of neighbors required to consider a conditional reliability as reliable}

\item{R}{A positive integer that indicates the number of repetitions for the bernoulli simulations.}

\item{seed.stab}{An integer used as argument by the set.seed() for offsetting the random number generator (Random integer by default). Only use if the study about the stability of predictions is required.}
}
\value{
A list of 7 objects is returned:
\item{seed}{The list of random number generator used. The first one is fixed by user or randomly chosen}
\item{nb.profil}{The number of profiles of covariates}
\item{conf.mat}{Global confusion matrix}
\item{res.prox}{A summary table related to the association measurements between Y and Z}
\item{res.grp}{A summary table related to the study of the proximity of Y and Z using groupings of levels}
\item{eff.neig}{A table which corresponds to a count of conditional probabilities according to the number of neighbors used in their computation (Only the first ten values)}
\item{res.stab}{A summary table related to the stability of the algorithm}
}
\description{
This function proposes post-process verifications after data fusion by Optimal Transportation algorithms
}
\details{
In a context of data fusion, where a same latent information is summarized via 2 specific variables Y and Z (two ordinal or nominal factors with different number of levels nY and nZ), stored in two distinct databases A and B respectively,
Optimal Transportation (OT) algorithms (\code{OUTCOME}, \code{R_OUTCOME}, \code{JOINT}, and \code{R_JOINT}, see the reference (2) for more details)
propose an original method for the recoding of Y in B and Z in A. The functions \code{OT_outcome} and \code{OT_joint} available in this package, make the use of these algorithms accessible,
and the function \code{verif_OT} uses their outputs by providing a set of tools (optional or not, depending on the choices done by user in input) to estimate:
\enumerate{
\item The association between Y and Z after recoding
\item The stability of the predictions proposed by the algorithm
}

A. PAIRWISE ASSOCIATION BETWEEN Y AND Z

The first step uses standard criterions (Cramer's V, chi square test of independence, Spearman coefficient of correlation) to evaluate associations between two ordinal variables in both databases or only in one.
When the argument \code{group.clss = TRUE}, these informations can be completed by those provided by the function \code{error_group} (available separately in the package), which is directly integrate in the function \code{verif_OT}.
Assuming that nY > nZ, this function studies all possible groupings of classes of the variable Y so that nY = nZ, in order to determine the grouping which will optimize the proximity between Y and Z.

B.STABILITY OF THE PREDICTIONS

These optional results are available by filling in the argument \code{stab.prob} (Provided by default).
Assuming that the missing information of Z in base A was predicted from an OT algorithm (the reasoning will be identical with the prediction of Y in B), the function \code{verif_OT} uses the conditional probabilities stored in the
object \code{estimatorZA} (see outputs of the functions \code{OT_outcome} and \code{OT_joint}) which contains the estimates of all the conditional probabilities of Z in A, kwowing a given profile of covariates x and kwowing a level of Y = y.
Each individual (or row) from A, can be so associated with a conditional probability \code{p_z,xy} and this latter is used to assess the stability of the algorithm in A.

Assuming that the stability of the algorithm in A is defined as its average ability to assign a same prediction of Z to individuals that have a same given profile of covariates x and a same given level of Y, for each individual we can draw one value from a Bernoulli distribution of parameter \code{p_z,xy} and
the average of these individual draws gives an estimation of the result.
Nevertheless, because estimates of conditional probabilities depends on the number of neighbors related to each profiles of covariates, some probabilities calculated from too small a number of individuals can be assumed to be unreliable and must be exclude of the final computations.
User can fix this minimal number of neighbor required in the estimation of a conditional probability to be remained in the final computation by using the argument \code{min.neigb} (By default all conditional probabilities are used).
In the case of the predictions of Z in A and Y in B are available, the function \code{verif_OT} provides in output, global and by databases results.
}
\examples{

### Example 1
#-----
# - Using the data simu_data
# - Studying the proximity between Y and Z using standard criterions
# - When Y and Z are predicted in B and A respectively
# - Using an outcome model (individual assignment with knn)
#-----
data(simu_data)
try1 = OT_outcome(simu_data, quanti = c(3,8), nominal = c(1,4:5,7), ordinal = c(2,6),
                 dist.choice = "G",percent.knn = 0.90, maxrelax = 0,
                 convert.num = 8, convert.clss = 3,
                 indiv.method = "sequential",which.DB = "BOTH",prox.dist = 0.30)

ver1 = verif_OT(try1); ver1


\dontrun{

### Example 2
#-----
# - Using the data simu_data
# - Studying the proximity between Y and Z using standard criterions and studying
#   associations by grouping levels of Z
# - When only Y is predicted in B
# - Using an outcome model (individual assignment with knn)
#-----

data(simu_data)
try2 = OT_outcome(simu_data, quanti = c(3,8), nominal = c(1,4:5,7), ordinal = c(2,6),
                 dist.choice = "G",percent.knn = 0.90, maxrelax = 0,
                 convert.num = 8, convert.clss = 3,
                 indiv.method = "sequential",which.DB = "B",prox.dist = 0.30)

ver2 = verif_OT(try2, group.clss = TRUE, ordinal = TRUE); ver2


### Example 3
#-----
# - Using the data simu_data
# - Studying the proximity between Y and Z using standard criterions and studying
#   associations by grouping levels of Z
# - Studying the stability of the conditional probabilities
# - When Y and Z are predicted in B and A respectively
# - Using an outcome model (individual assignment with knn)
#-----

ver3 = verif_OT(try2, group.clss = TRUE, ordinal = TRUE, stab.prob = TRUE, min.neigb = 5,
       seed.stab = 991891); ver3

}

}
\references{
\enumerate{
\item Gares V, Dimeglio C, Guernec G, Fantin F, Lepage B, Korosok MR, savy N (2019). On the use of optimal transportation theory to recode variables and application to database merging. The International Journal of Biostatistics.
Volume 16, Issue 1, 20180106, eISSN 1557-4679 | \url{https://doi.org/10.1515/ijb-2018-0106}
\item Gares V, Omer J. Regularized optimal transport of covariates and outcomes in datarecoding(2019).hal-02123109 \url{https://hal.archives-ouvertes.fr/hal-02123109/document}
}
}
\seealso{
\code{\link{OT_outcome}}, \code{\link{OT_joint}}, \code{\link{proxim_dist}}, \code{\link{error_group}}
}
\author{
Gregory Guernec
\email{otrecod.pkg@gmail.com}
}
