% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_group.r
\name{error_group}
\alias{error_group}
\title{error_group()}
\usage{
error_group(REF, Z, ord = TRUE)
}
\arguments{
\item{REF}{A factor with a reference number of levels}

\item{Z}{A factor with a number of levels greater than the number of levels of the reference}

\item{ord}{A boolean. If TRUE, only the subsequent levels of Z can be grouped together}
}
\value{
A data.frame ot 2 columns:
        The first column enumerates all possible groups of modalities of Z to obtain the same number of levels as the reference
        The second columns gives the corresponding rate error of reclassification
}
\description{
error_group()
}
\examples{
Z1 = as.factor(sample(1:3,50,replace = TRUE)); length(Z1)
Z3 = as.factor(sample(1:2,50,replace = TRUE)); length(Z3)
Z2 = as.factor(sample(c("A","B","C","D"),50, replace = TRUE)); length(Z2)
Z4 = as.factor(sample(c("A","B","C","D","E"),50, replace = TRUE)); length(Z4)
error_group(Z1,Z4)
error_group(Z3,Z1,FALSE)

}
