% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata_nd.R
\name{gendata_ND}
\alias{gendata_ND}
\title{SIMULATION DE DONNEES POUR APPROCHE NON DETERMINISTE
-> Cette fonction genere une BDD de la forme [Base|Y1|Y2|X1|X2|X3] avec les
   covariables X1, X2 categorielles ? k1 et K2 modalites resp., et X3 continue}
\usage{
gendata_ND(n1, k, cormat, px1c, px2c, r2, py1c, py2c, valY1, valY2)
}
\arguments{
\item{n1}{: Nombre de patients dans la base 1}

\item{k}{: Coefficient de proportionnalite entre les 2 bases}

\item{cormat}{: vecteur contenant les correlations: rho = cor(X1,X2),
delta = cor(X1,X3), mu = cor(X2,X3)}

\item{px1c}{: Proportion de patients dans chaque classe de X1}

\item{px2c}{: Proportion de patients dans chaque classe de X2}

\item{r2}{: R^2 de la regression lineaire Ycontinue = f(X1,X2,X3)}

\item{py1c}{: Proportion de patients dans chaque classe de Y en base 1}

\item{py2c}{: Proportion de patients dans chaque classe de Y en base 2}

\item{valY1}{: Les labels de chaque classe pour Y cat?gorielle en base 1}

\item{valY2}{: Les labels de chaque classe pour Y cat?gorielle en base 2}

\item{rho}{: Correlation entre X1 et X2 (continues)}

\item{delta}{: Correlation entre X1 et X3 (continues)}

\item{mu}{: Correlation entre X2 et X3 (continues)}
}
\value{
missing documentation
}
\description{
Quel que soit le nombre de classes de X1 et X2, elles 
sont au format numerique en sortie
}
