% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.R
\name{imp_MICE}
\alias{imp_MICE}
\title{MICE - Liste des methodes possible pour l'imputation:}
\usage{
imp_MICE(datessai, R_mice, meth, calc)
}
\arguments{
\item{datessai}{Jeu de donn?es AVEC les NA sur Y1 et Y2
requiert que les outcomes s'appellent "Y1" et "Y2"}

\item{R_mice}{Nbr de r?p?titions pour MICE}

\item{meth}{Vecteur des m?thodes d'imputation ? appliquer ?
chaque variable et covariable selon son type: quanti ou facteur "pmm" si toutes quantis
"polyreg" si qualis (facteur) ? >= 2 modalit?s}

\item{calc}{Doit-on moyenner les valeurs imput?es ou conserver les plus
fr?quentes ? ="MOY" ou ="FREQ"}
}
\value{
un resultat
}
\description{
pmm         :  Predictive mean matching (any)
norm        : Bayesian linear regression (numeric)
norm.nob    : Linear regression ignoring model error (numeric)
norm.boot   : Linear regression using bootstrap (numeric)
norm.predict: Linear regression, predicted values (numeric)
mean        : Unconditional mean imputation (numeric)
2l.norm     : Two-level normal imputation (numeric)
2l.pan      : Two-level normal imputation using pan (numeric
2lonly.mean : Imputation at level-2 of the class mean (numeric)
2lonly.norm : Imputation at level-2 by Bayesian linear regression (numeric)
2lonly.pmm  : Imputation at level-2 by Predictive mean matching (any)
quadratic   : Imputation of quadratic terms (numeric)
logreg      : Logistic regression (factor, 2 levels)
logreg.boot : Logistic regression with bootstrap
polyreg     : Polytomous logistic regression (factor, >= 2 levels)
polr        : Proportional odds model (ordered, >=2 levels)
lda         : Linear discriminant analysis (factor, >= 2 categories)
cart        : Classification and regression trees (any)
rf          : Random forest imputations (any)
ri          : Random indicator method for nonignorable data (numeric)
sample      : Random sample from the observed values (any)
fastpmm     : Experimental: Fast predictive mean matching using C++ (any)
}
