% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_grp_optimal.r
\name{indiv_grp_optimal}
\alias{indiv_grp_optimal}
\title{indiv_grp_optimal()}
\usage{
indiv_grp_optimal(proxim, jointprobaA, jointprobaB, percent_closest = 1)
}
\arguments{
\item{proxim}{An object corresponding to the output of the \code{\link{proxim_dist}} function}

\item{jointprobaA}{A matrix which number of columns equals the number of modalities of the target variable Y in database A, and which number of rows equals the number of modalities of Z in database B. It gives an estimation of the joint probability (Y,Z) in DB A.
The sum of cells of this matrix must be equal to 1}

\item{jointprobaB}{A matrix which number of columns equals the number of modalities of the target variable Y in database A, and which number of rows equals the number of modalities of Z in database B. It gives an estimation of the joint probability (Y,Z) in DB B.
The sum of cells of this matrix must be equal to 1}

\item{percent_closest}{A value between 0 and 1 (by default) corresponding to the desired \code{percent closest} indivisuals taken in the computation of the distances}
}
\value{
A list of two vectors of numeric values:
\describe{
\item{YAtrans}{A vector corresponding to the predicted values of Y in database B using the Optimal Transportation theory}
\item{YBtrans}{A vector corresponding to the predicted values of Z in database A using the Optimal Transportation theory}
}
}
\description{
Assuming that Y and Z summarize a same information encoded in two distinct forms stored in two distinct databases A and B.
This function is an alternative to the \code{\link{indiv_grp_closest}} approach for the individual predictions of Y en B and Z in A.
It solves an optimization problem to get the individual transport that minimizes total distance while satisfying the joint probability computed by the model.
See section 5.3 (p24) of the reference article 2 for more details about this method.
}
\details{
As for \code{\link{indiv_grp_closest}}, the \code{indiv_grp_optimal} function is directly implemented in the \code{OT} and requires the prior use of code{\link{proxim_dist}} and \code{\link{avg_dist_closest}} for running.
}
\references{
# Article 1:
Gares V, Dimeglio C, Guernec G, Fantin F, Lepage B, Korosok MR, savy N (2019). On the use of optimal transportation theory to recode variables and application to database merging. The International Journal of Biostatistics.
0, 20180106 (2019) | \url{https://doi.org/10.1515/ijb-2018-0106}

# Article 2:
Gares V, Omer J. Regularized optimal transport of covariates and outcomes in datarecoding(2019).hal-02123109 \url{https://hal.archives-ouvertes.fr/hal-02123109/document}
}
\seealso{
\code{\link{proxim_dist}}, \code{\link{avg_dist_closest}}, \code{\link{indiv_grp_closest}}
}
\author{
Gregory Guernec, Valerie Gares, Jeremy Omer
\email{gregory.guernec@inserm.fr}
}
