% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_grp_optimal.r
\name{indiv_grp_optimal}
\alias{indiv_grp_optimal}
\title{indiv_grp_optimal()}
\usage{
indiv_grp_optimal(
  proxim,
  jointprobaA,
  jointprobaB,
  percent_closest = 1,
  which.DB = "BOTH"
)
}
\arguments{
\item{proxim}{An object corresponding to the output of the function \code{\link{proxim_dist}}}

\item{jointprobaA}{A matrix whose number of columns is equal to the number of modalities of the target variable Y in database A, and which number of rows is equal to the number of modalities of Z in database B. It gives an estimation of the joint probability (Y,Z) in the database A.
The sum of cells of this matrix must be equal to 1}

\item{jointprobaB}{A matrix whose number of columns is equal to the number of modalities of the target variable Y in database A, and which number of rows is equal to the number of modalities of Z in database B. It gives an estimation of the joint probability (Y,Z) in the database B.
The sum of cells of this matrix must be equal to 1}

\item{percent_closest}{A value between 0 and 1 (by default) corresponding to the fixed \code{percent closest} of individuals used in the computation of the average distances}

\item{which.DB}{A character string (with quotes) that indicates which individual predictions compute: only the individual predictions of Y in B ("B"), only those of Z in A ("A") or the both ("BOTH" by default)}
}
\value{
A list of two vectors of numeric values:
\item{YAtrans}{A vector corresponding to the predicted values of Y in database B (numeric form) using the Optimal Transportation theory}
\item{ZBtrans}{A vector corresponding to the predicted values of Z in database A (numeric form) using the Optimal Transportation theory}
}
\description{
This function assigns individual predictions related to recoding problems of data fusion by solving a linear optimization problem
}
\details{
A. THE RECODING PROBLEM IN DATA FUSION

Assuming that Y and Z are two variables that summarize a same latent information in two separate (no overlapping rows) databases A and B respectively,
so that Y and Z are never jointly observed in A and B. Assuming also that A and B share a subset of common covariates X of any types (but same encodings in A and B)
completed or not. Integrating these two databases often requires to solve the recoding problem observed between Y and Z by creating an unique database where
the missing information of Y and Z is completed.


B. DESCRIPTION OF THE FUNCTION

The function \code{indiv_grp_optimal} is an intermediate function used in the implementation of a model called {OUTCOME} (and its enrichment {R-OUTCOME}, see the reference (2) for more details) dedicated to the solving of recoding problems in data fusion using Optimal Transportation theory.
The model is implemented in the function \code{\link{OT_outcome}} which integrates the function \code{indiv_grp_optimal} in its syntax as a possible second step of the algorithm.
The function \code{indiv_grp_optimal} can nevertheless be used separately provided that the argument \code{proxim} receives an output object of the function \code{\link{proxim_dist}}.
This latter is available in the package and is so directly usable beforehand.

The function \code{indiv_grp_optimal} constitutes an alternative method to the nearest neighbor procedure implemented in the function \code{\link{indiv_grp_closest}}. Its objective consists in getting individual transports from grouped joint probabilities
by directly targetting the solution of recoding problems without using a nearest neighbor approach which constraints to some arbitrary decisions.
As for the function \code{\link{indiv_grp_closest}}, assuming that the objective consists in the prediction of Z in the database A, the first step of the algorithm related to \code{OUTCOME} provides an estimate of gamma which can be interpreted as an estimation of the joint distribution (Y,Z) in A.
Rather than using a nearest neighbor approach to provide individual predictions,the function \code{indiv_grp_optimal} solves an optimization problem using the simplex algorithm which searches for the individual predictions of Z that minimize the computed total distance observed satisfying the joint probability distribution estimated in the first part.
More details about the theory related to the solving of this optimization problem is described in the section 5.3 of (2).

Obviously, this algorithm  runs in the same way for the prediction of Y in the database B.
The function \code{indiv_grp_optimal} integrates in its syntax the function \code{\link{avg_dist_closest}} and the related argument \code{percent_closest} is identical in the two functions.
Thus, when computing average distances between an individual i and a subset of individuals assigned to a same level of Y or Z is required, user can decide if all individuals from the subset of interest can participate to the computation (\code{percent_closest}=1) or only a fixed part p (<1) corresponding to the closest neighbors of i (in this case \code{percent_closest} = p).

The arguments \code{jointprobaA} and \code{jointprobaB} can be seen as cost matrices (sum of cells must be equal to 1) that correspond to estimations of the joint distributions of (Y,Z) in A and B respectively.
}
\examples{

### Example using The Euclidean distance on a complete database
# For this example we keep only 200 rows:

data(tab_test)
tab_test2 = tab_test[c(1:80,5001:5080),]; dim(tab_test2)

# Adding NAs in Y1 and Y2
tab_test2[tab_test2$ident == 2, 2] = NA
tab_test2[tab_test2$ident == 1, 3] = NA

# Because all covariates are ordered in numeric form,
# the transfo_dist function is not required here

res3      = proxim_dist(tab_test2,norm = "M")

#' ### Y(Y1) and Z(Y2) are a same variable encoded in 2 different forms:
### 4 levels for Y1 and 3 levels for Y2
### ... Stored in two distinct databases, A and B, respectively
### The marginal distribution of Y in B is unknown,
### as the marginal distribution of Z in A ...

# Assuming that the following matrix called transport symbolizes
# an estimation of the joint distribution L(Y,Z) ...
# Note that, in reality this distribution is UNKNOWN and is
# estimated in the OT function by resolving the optimization problem.

# By supposing:

val_trans  = c(0.275,0.115,0,0,0,0.085,0.165,0,0,0,0.095,0.265)
transport2 = matrix(val_trans,ncol = 3,byrow = FALSE)

# Getting the individual predictions of Z in A (only)
# by computing average distances on 90\% of the nearest neighbors of
# each modality of Z in B
res4      = indiv_grp_optimal(res3,jointprobaA = transport2,
            jointprobaB = transport2, percent_closest= 0.90,
            which.DB = "A")

\dontrun{
### Example 2 using The Manhattan distance with incomplete covariates
data(simu_data)

try1 = transfo_dist(simu_data,quanti = c(3,8), nominal = c(1,4:5,7),
                    ordinal = c(2,6), logic = NULL, prep_choice = "M")
res1 = proxim_dist(try1,norm = "M")


### Y and Z are a same variable encoded in 2 different forms:
### (3 levels for Y and 5 levels for Z)
### ... Stored in two distinct databases, A and B, respectively
### The marginal distribution of Y in B is unknown,
### as the marginal distribution of Z in A ...

# By supposing that the following matrix called transport symbolizes
# an estimation of the joint distribution L(Y,Z) ...
# Note that, in reality this distribution is UNKNOWN and is
# estimated in the OT function by resolving an optimisation problem.

transport1 = matrix(c(0,0.35285714,0,0.09142857,0,0.03571429,
                      0,0,0.08285714,0,0.07857143,0.03142857,
                      0.32714286,0,0),ncol = 5,byrow = FALSE)


# The predicted values of Y in database B and Z in
database A are stored in the following object:

res2 = indiv_grp_optimal(res1,jointprobaA = transport1,
                         jointprobaB = transport1,
                         percent_closest= 0.90)                               )
summary(res2)
}

}
\references{
\enumerate{
\item Gares V, Dimeglio C, Guernec G, Fantin F, Lepage B, Korosok MR, savy N (2019). On the use of optimal transportation theory to recode variables and application to database merging. The International Journal of Biostatistics.
Volume 16, Issue 1, 20180106, eISSN 1557-4679 | \url{https://doi.org/10.1515/ijb-2018-0106}
\item Gares V, Omer J (2020) Regularized optimal transport of covariates and outcomes in data recoding. Journal of the American Statistical Association, DOI: 10.1080/01621459.2020.1775615
}
}
\seealso{
\code{\link{proxim_dist}}, \code{\link{avg_dist_closest}}, \code{\link{indiv_grp_closest}}
}
\author{
Gregory Guernec, Valerie Gares, Jeremy Omer
\email{otrecod.pkg@gmail.com}
}
