% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_grp_optimal.r
\name{indiv_grp_optimal}
\alias{indiv_grp_optimal}
\title{indiv_grp_optimal()}
\usage{
indiv_grp_optimal(
  proxim,
  jointprobaA,
  jointprobaB,
  percent_closest = 1,
  which.DB = "BOTH"
)
}
\arguments{
\item{proxim}{An object corresponding to the output of the function \code{\link{proxim_dist}}}

\item{jointprobaA}{A matrix whose number of columns equals the number of modalities of the target variable Y in database A, and which number of rows equals the number of modalities of Z in database B. It gives an estimation of the joint probability (Y,Z) in the database A.
The sum of cells of this matrix must be equal to 1}

\item{jointprobaB}{A matrix whose number of columns equals the number of modalities of the target variable Y in database A, and which number of rows equals the number of modalities of Z in database B. It gives an estimation of the joint probability (Y,Z) in the database B.
The sum of cells of this matrix must be equal to 1}

\item{percent_closest}{A value between 0 and 1 (by default) corresponding to the fixed \code{percent closest} of individuals taken in the computation of the average distances}

\item{which.DB}{A character string (with quotes) that indicates which individual predictions compute: Only the individual predictions of Y in B ("B"), only those of Z in A ("A") or the both ("BOTH" by default)}
}
\value{
A list of two vectors of numeric values:
\item{YAtrans}{A vector corresponding to the predicted values of Y in database B (numeric form) using the Optimal Transportation theory}
\item{ZBtrans}{A vector corresponding to the predicted values of Z in database A (numeric form) using the Optimal Transportation theory}
}
\description{
This function assigns individual predictions related to recoding problems of data fusion by solving a linear optimization problem
}
\details{
A. THE RECODING PROBLEM IN DATA FUSION

When two databases are constructed from heterogeneous sources, it is not usual that two different encodings can be used for a same target outcome.
Assuming that Y and Z are these 2 variables that summarize a same latent information in 2 separated (no overlapping rows) databases A and B respectively,
so that Y and Z are never jointly observed in A and B. Assuming also that A and B share a subset of common covariates X of any types (but same encodings in A and B)
complete or not. Integrating these two databases often requires to solve the recoding problem observed between Y in Z by creating an unique database where
the missing information of Y and Z is completed.


B. DESCRIPTION OF THE FUNCTION

The function \code{indiv_grp_optimal} is an intermediate function used in the implementation of an original algorithm dedicated to the solving of recoding problems in data fusion using Optimal Transportation theory (For more details about the algorithm, see the the models
\code{OUTCOME},\code{R_OUTCOME},\code{JOINT} and \code{R_JOINT}, described in the references (1) and (2)).
\code{\link{indiv_grp_optimal}} is so directly implemented in the \code{OT_outcome} and \code{OT_joint} functions but can also be used separately.

This function constitutes an alternative method to the nearest neighbor procedure implemented in the function \code{\link{indiv_grp_closest}}. Its objective consists in getting individual transports from grouped joint probabilities
by directly targetting the solution of recoding problems without using a nearest neighbor approach which constraints to some arbitrary decisions.
To avoid this, the approach implemented rather solves an optimization problem using the simplex algorithm to get the individual transport that minimizes total distance while satisfying the joint probability distribution provides by the arguments \code{jointprobaA} and \code{jointprobaB}.
More details about the theory related to the solving of this optimization problem is described in the section 5.3 of (2).

As for \code{\link{indiv_grp_closest}}:
\itemize{
\item The function \code{indiv_grp_optimal} requires the use of code{\link{proxim_dist}} and \code{\link{avg_dist_closest}} for running.
Nevertheless, if the second one is directly integrated in the function, the specific output of the first one stays required in input of this latter.
\item The arguments \code{jointprobaA} and \code{jointprobaB} are here transport matrices (where the sum of cells must be equal to 1) that correspond to estimations of gamma, the map that respectively pushes the distribution of Y forward to the distribution of Z and inversely.
}
All individuals can participate to the computation of the average distances (\code{percent_closest}=1) or only a fixed part p corresponding to the closest neighbors of each individual from each modalities of the outcomes (in this case \code{percent_closest} < p).
}
\examples{

### Example using The Euclidean distance on a complete database
# For this example we keep only 200 rows:

data(tab_test)
tab_test2 = tab_test[c(1:80,5001:5080),]; dim(tab_test2)

# Adding NAs in Y1 and Y2
tab_test2[tab_test2$ident == 2,2] = NA
tab_test2[tab_test2$ident == 1,3] = NA

# Because all covariates are ordered in numeric form,
# the transfo_dist function is not required here

res3      = proxim_dist(tab_test2,norm = "M")

#' ### Y(Y1) and Z(Y2) are a same variable encoded in 2 different forms:
### 4 levels for Y1 and 3 levels for Y2
### ... Stored in two distinct databases, A and B, respectively
### The marginal distribution of Y1 in B is unknown,
### as the marginal distribution of Z2 in A ...

# By supposing that the following matrix called transport symbolizes
# an estimation of the joint distribution L(Y,Z) ...
# Note that, in reality this distribution is UNKNOWN and is
# estimated in the OT function by resolving an optimisation problem.

# By supposing:

val_trans  = c(0.275,0.115,0,0,0,0.085,0.165,0,0,0,0.095,0.265)
transport2 = matrix(val_trans,ncol = 3,byrow = FALSE)

# Getting the individual predictions of Z in A (only)
# by computing average distances on 90\% of the nearest neighbors of
# each modality of Z in B
res4      = indiv_grp_optimal(res3,jointprobaA = transport2,
            jointprobaB = transport2, percent_closest= 0.90,
            which.DB = "A")

\dontrun{
### Example 2 using The Manhattan distance with incomplete covariates
data(simu_data)

try1 = transfo_dist(simu_data,quanti = c(3,8), nominal = c(1,4:5,7),
                    ordinal = c(2,6), logic = NULL, prep_choice = "M")
res1 = proxim_dist(try1,norm = "M")


### Y and Z are a same variable encoded in 2 different forms:
### (3 levels for Y and 5 levels for Z)
### ... Stored in two distinct databases, A and B, respectively
### The marginal distribution of Y in B is unknown,
### as the marginal distribution of Z in A ...

# By supposing that the following matrix called transport symbolizes
# an estimation of the joint distribution L(Y,Z) ...
# Note that, in reality this distribution is UNKNOWN and is
# estimated in the OT function by resolving an optimisation problem.

transport1 = matrix(c(0,0.35285714,0,0.09142857,0,0.03571429,
                      0,0,0.08285714,0,0.07857143,0.03142857,
                      0.32714286,0,0),ncol = 5,byrow = FALSE)


# The affectation of the predicted values of Y in database B and Z in
database A are stored in the following object:

res2 = indiv_grp_optimal(res1,jointprobaA = transport1,
                         jointprobaB = transport1,
                         percent_closest= 0.90)                               )
summary(res2)
}

}
\references{
\enumerate{
\item Gares V, Dimeglio C, Guernec G, Fantin F, Lepage B, Korosok MR, savy N (2019). On the use of optimal transportation theory to recode variables and application to database merging. The International Journal of Biostatistics.
Volume 16, Issue 1, 20180106, eISSN 1557-4679 | \url{https://doi.org/10.1515/ijb-2018-0106}
\item Gares V, Omer J (2020) Regularized optimal transport of covariates and outcomes in data recoding. Journal of the American Statistical Association, DOI: 10.1080/01621459.2020.1775615
}
}
\seealso{
\code{\link{proxim_dist}}, \code{\link{avg_dist_closest}}, \code{\link{indiv_grp_closest}}
}
\author{
Gregory Guernec, Valerie Gares, Jeremy Omer
\email{otrecod.pkg@gmail.com}
}
