% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_from_group_closest.r
\name{individual_from_group_closest}
\alias{individual_from_group_closest}
\title{Title}
\usage{
individual_from_group_closest(inst, jointprobaA, jointprobaB,
  percent_closest = 1)
}
\arguments{
\item{inst}{}

\item{jointprobaA}{}

\item{jointprobaB}{}

\item{percent_closest}{}
}
\value{
list
}
\description{
Title
}
