% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_from_group_optimal.r
\name{individual_from_group_optimal}
\alias{individual_from_group_optimal}
\title{individual_from_group_optimal(inst, jointprobaA, jointprobaB, percent_closest=1.0)}
\usage{
individual_from_group_optimal(inst, jointprobaA, jointprobaB,
  percent_closest = 1)
}
\arguments{
\item{inst}{todo list}

\item{jointprobaA}{todo list}

\item{jointprobaB}{todo list}

\item{percent_closest}{todo list}
}
\value{
todo list
}
\description{
individual_from_group_optimal(inst, jointprobaA, jointprobaB, percent_closest=1.0)
}
