% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OT.r
\name{OT}
\alias{OT}
\title{Title}
\usage{
OT(datab, percent_c = 1, maxrelax = 0, norm = 1, indiv_method,
  full_disp = FALSE, solver_disp = FALSE)
}
\arguments{
\item{datab}{}

\item{percent_c}{}

\item{maxrelax}{}

\item{norm}{}

\item{indiv_method}{}

\item{full_disp}{}

\item{solver_disp}{}
}
\value{
list with time and solution
}
\description{
Title
}
