% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OT_joint.r
\name{OT_joint}
\alias{OT_joint}
\title{OT_joint(datab, 
maxrelax=0.0, 
lambda_reg = 0.0, 
percent_clos = 0.2, 
norm = 1, 
aggregate_tol=0.5, 
full_disp = FALSE, 
solver_disp = FALSE)}
\usage{
OT_joint(datab, maxrelax = 0, lambda_reg = 0, percent_clos = 0.2,
  norm = 1, aggregate_tol = 0.5, full_disp = FALSE,
  solver_disp = FALSE)
}
\arguments{
\item{datab}{todo list}

\item{maxrelax}{todo list}

\item{lambda_reg}{todo list}

\item{percent_clos}{todo list}

\item{norm}{todo list}

\item{aggregate_tol}{todo list}

\item{full_disp}{todo list}

\item{solver_disp}{todo list}
}
\value{
List with time and solution
}
\description{
OT_joint(datab, 
maxrelax=0.0, 
lambda_reg = 0.0, 
percent_clos = 0.2, 
norm = 1, 
aggregate_tol=0.5, 
full_disp = FALSE, 
solver_disp = FALSE)
}
