% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OT_joint.r
\name{OT_joint}
\alias{OT_joint}
\title{Title}
\usage{
OT_joint(datab, maxrelax = 0, lambda_reg = 0, percent_clos = 0.2,
  norm = 1, aggregate_tol = 0.5, full_disp = FALSE,
  solver_disp = FALSE)
}
\arguments{
\item{datab}{}

\item{maxrelax}{}

\item{lambda_reg}{}

\item{percent_clos}{}

\item{norm}{}

\item{aggregate_tol}{}

\item{full_disp}{}

\item{solver_disp}{}
}
\value{
List with time and solution
}
\description{
Title
}
