% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ot_joint.r
\name{OT_joint}
\alias{OT_joint}
\title{OT_joint()}
\usage{
OT_joint(
  datab,
  index_DB_Y_Z = 1:3,
  nominal = NULL,
  ordinal = NULL,
  logic = NULL,
  convert.num = NULL,
  convert.clss = NULL,
  dist.choice = "E",
  percent.knn = 1,
  maxrelax = 0,
  lambda.reg = 0,
  prox.dist = 0.8,
  prox.X = 0.8,
  which.DB = "BOTH"
)
}
\arguments{
\item{datab}{A data.frame with at least four columns sorted in a non-specific order. One column must be a column of identifier for the two databases, where the names of the two databases
must be ranked in ascending order (By example: 1 for the top database and 2 for the database from below, or more logically here A and B  ...But NOT B and A!). One column (\eqn{Y} here but other names are allowed)
must correspond to the target variable related to the information of interest to merge with its specific encoding in the database A (corresponding encoding should be so missing in the database B). In the same way,
one column (\eqn{Z} here) corresponds to the second target variable with its specific encoding in the database B (corresponding encoding should be so missing in the database A).
Finally, the input database must have at least one shared covariate with same encoding in A and B. Please notice that, if your data.frame has only four columns, that is to say with only one covariate, if this latter has NA, and unless user
has previously imputed the corresponding missing information, the OT algorithm will only run complete cases.}

\item{index_DB_Y_Z}{A vector of three column indexes. The first index must correspond to the index of the databases identifier column (DB identifier). The second index must correspond
to the index of the target variable in the first database (A) while the third index corresponds to the index of column related to the target variable in the second database (B).}

\item{nominal}{A vector of indexes that corresponds to the column indexes of all the nominal (not ordered) variables (DB identification and target variables included if it is the case for them).}

\item{ordinal}{A vector of indexes that corresponds to the column indexes of all the ordinal variables (DB identification and target variables included if it is the case for them).}

\item{logic}{A vector of indexes that corresponds to the column indexes of all the boolean variables of the data.frame.}

\item{convert.num}{Indexes of the continuous (quantitative) variables. They will be automatically converted in ordered factors. By default, no continuous variables is assumed in the database.}

\item{convert.clss}{A vector indicating for each continuous variable to convert, the corresponding desired number of levels. If the length of the argument \code{convert_num} exceeds 1 while the length of \code{convert_clss} equals 1 (only one integer),
each discretization will count the same number of levels.}

\item{dist.choice}{A character (with quotes) corresponding to the distance function chosen between: the euclidean distance ("E", by default), the Manhattan distance ("M"),
the Gower distance ("G"), and the Hamming distance ("H") for binary covariates only.}

\item{percent.knn}{Percent of closest neighbors taken in the computation of the cost matrix.}

\item{maxrelax}{The maximum percentage of deviation from expected probability masses. It must be equal to 0 (default value) for the \code{JOINT} model, and equal to a strictly positive value for the R-JOINT model}

\item{lambda.reg}{A coefficient measuring the importance of the regularization term. In the related reference (2), it corresponds to the \code{R-JOINT} model for a value other than 0 (default value))}

\item{prox.dist}{A percentage (betwen 0 and 1) used to calculate the distance threshold below which an individual (a row) is considered as a neighbor of a given profile of covariates.}

\item{prox.X}{A percentage (betwen 0 and 1) used to calculate the distance threshold below which two covariates' profiles are supposed as neighbors.
If \code{prox.X = 1}, all profiles are considered as neighbors.}

\item{which.DB}{A character indicating the database to complete ("BOTH" by default, for the prediction of \eqn{Y} and \eqn{Z} in the two databases), "A" only for the imputation of \eqn{Z} in A, "B" only for the imputation of \eqn{Y} in B.}
}
\value{
A list of 7 elements containing:
    \item{time_exe}{Running time of the function}
    \item{gamma_A}{Estimation of \eqn{\gamma} for the completion of A. A cost matrix that corresponds to the joint distribution of \eqn{(Y^A,Z^A,X^A)}}
    \item{gamma_B}{Estimation of \eqn{\gamma} for the completion of B. A cost matrix that corresponds to the joint distribution of \eqn{(Y^B,Z^B,X^B)}}
    \item{profile}{A data.frame that gives all details about the remaining P profiles of covariates. These informations can be linked to the \code{estimatorZA} and the \code{estimatorYB} objects for a better interpretation of the results}
    \item{res_prox}{The outputs of the function \code{proxim_dist}}
    \item{estimatorZA}{An array that corresponds to estimates of the probability distribution of \eqn{Z} conditional to \eqn{X} and \eqn{Y} in database A. The number of rows of each table corresponds to the total number of profiles of covariates.
    The number of columns of each table corresponds to the number of levels of \eqn{Y}. The row names of each table corresponds to the values of the covariates sorted by order of appearance in the merged database. The third element of the array is the possible level of \eqn{Z}.}
    \item{estimatorYB}{An array that corresponds to estimates of the probability distribution of \eqn{Y} conditional to \eqn{X} and \eqn{Z} in database B. The number of rows of each table corresponds to the total number of profiles of covariates.
    The number of columns of each table corresponds to the number of levels of \eqn{Z}. The row names of each table corresponds to the values of the covariates sorted by order of appearance in the merged database. The third element of the array is the possible level of \eqn{Y}.}
    \item{DATA1_OT}{database A with imputed individual prediction on \eqn{Z} using OT}
    \item{DATA2_OT}{database B with imputed individual prediction on \eqn{Y} using OT}
}
\description{
The function \code{OT_joint} integrates two models called (\code{JOINT}) and (\code{R-JOINT}) dedicated to the solving of recoding problems in data integration
using optimal transportation of the joint distribution of outcomes and covariates.
}
\details{
A. THE RECODING PROBLEM IN DATA FUSION

Assuming that \eqn{Y} and \eqn{Z} are two variables that summarize a same latent information in two separate (no overlapping rows) databases A and B respectively,
so that \eqn{Y} and \eqn{Z} are never jointly observed in A and B. Assuming also that A and B share a subset of common covariates X of any types (with same encodings in A and B)
completed or not. Integrating these two databases often requires to solve the recoding problem observed between \eqn{Y} and \eqn{Z} by creating an unique database where
the missing information of \eqn{Y} and/or \eqn{Z} is fully completed.


B. INFORMATIONS ABOUT THE ALGORITHM

As with the function \code{\link{OT_outcome}}, the function \code{OT_joint} provides a solution to the recoding problem by proposing an
application of optimal transportation which aims is to search for a bijective mapping between the joint distributions of \eqn{(Y,X)} and \eqn{(Z,X)} in A and B (see (2) for more details).
The principle of the algorithm is also based on the resolution of an optimization problem, which provides a \eqn{\gamma} solution (as called in (1) and (2)) which is an estimation
of the joint distribution of \eqn{(X,Y,Z)} according to the database to complete (see the argument \code{which.DB} for the choice of the database). While the models \code{OUTCOME} and \code{R_OUTCOME} integrated in
the function \code{\link{OT_outcome}} require post-treatment steps to provide individual predictions, the algorithm \code{JOINT} direcly uses estimations of the conditional distributions \eqn{(Y|Z,X)} in B and
\eqn{(Z|Y,X)} in A to predict the corresponding incomplete individuals informations of \eqn{Y} and/or \eqn{Z} respectively.
This algorithm supposes that the conditional distribution \eqn{(Y|X)} must be identical in A and B. Respectively, \eqn{(Z|X)} is supposed identical in A and B.
Estimations a posteriori of conditional probabilities \eqn{P[Y|X,Z]} and \eqn{P[Z|X,Y]} are available by profiles of covariates in output (See the objects \code{estimatorYB} and \code{estimatorZA}).
Estimations of \eqn{\gamma} are also available according to the chosen transport distributions (See the arguments \eqn{\gamma_A} and \eqn{\gamma_B}).

The model \code{R-JOINT} gathers enrichments of the model \code{JOINT} and is also available via the function \code{OT_joint} It allows users to add a relaxation term in the algorithm to relax distributional assumptions (\code{maxrelax}>0),
and (or) add also a positive regularization term (\code{lamdba.reg}>0) expressing that the transportation map does not vary to quickly with respect of covariates \eqn{X}.
Is is suggested to users to calibrate these two parameters a posteriori by studying the stability of the individual predictions in output.


C. TYPE OF REQUIRED INPUT DATABASE

The input database is a data.frame that must be saved in a specific form by users:
\itemize{
\item Two overlayed databases containing a common column of databases' identifiers (A and B, 1 or 2, by examples, encoded in numeric or factor form)
\item A column corresponding to the target variable with its specific encoding in A (For example a factor \eqn{Y} encoded in \eqn{n_Y} levels, ordered or not, with NAs in the corresponding rows of B)
\item A column corresponding to another target outcome summarizing the same latent information with its specific endoded in B (By example a factor \eqn{Z} in \eqn{n_Z} levels, with NAs in rows of A)
\item The order of the variables in the database have no importance but the indexes of the columns related to the 3rd columns previously described (ie ID, \eqn{Y} and \eqn{Z}) must be rigorously specified
in the argument \code{index_DB_Y_Z}.
\item A set of shared common categorical covariates (at least one but more is recommended) complete or not (provided that the number of covariates exceeds 1) is required. On the contrary to the
function \code{OT_outcome}, please notice, that the function \code{OT_joint} does not accept continuous covariates therefore these latters will have to be categorized beforehand or using the provided input process (see \code{quanti}).
}
The function \code{\link{merge_dbs}} is available in this package to assist user in the preparation of their databases, so please, do not hesitate to use it beforehand if necessary.

Remarks about the outcomes:
\itemize{
\item A target outcome can be categorical, in factor, ordered or not, discrete (with a finite number of values ONLY) but, notice that, if they are stored in numeric they will be automatically converted in ordered factors.
\item If a target outcome is incomplete, the corresponding rows will be automatically dropped during the execution of the function.
}
The type of each variables (including ID, \eqn{Y} and \eqn{Z}) of the database must be rigorously specified, in one of the four arguments \code{quanti}, \code{nominal}, \code{ordinal} and \code{logic}.


D. TRANSFORMATIONS OF CONTINUOUS COVARIATES

Continuous covariates with infinite numbers of values will have to be categorized beforehand their inclusions in the function.
To assist users in this task, the function \code{OT_joint} integrates in its syntax a process dedicated to the categorization of continuous covariates. For this, it is necessary to rigorously fill in
the arguments \code{quanti} and \code{convert.clss}.
The first one informs about the indexes in database of the continuous variables to transform in ordered factor while the second one specifies the corresponding number of desired balanced levels (for unbalanced levels, users must do transformations by themselves).
Therefore \code{convert.num} and \code{convert.clss} must be vectors of same length, but if the length of \code{quanti} exceeds 1, while the length of \code{convert.clss} is 1, then, by default, all the covariates to convert will have the same number of classes (transformation by quantiles),
that corresponds to the value specified in the argument \code{convert.clss}.
Please notice that only covariates can be transformed (not outcomes) and missing informations are not taken into account for the transformations.
Moreover, all the indexes informed in the argument \code{convert.num} must also be informed in the argument \code{quanti}.
Finally, it is suggested to declare all discrete covariates as ordinal factors using the argument \code{ordinal}.


E. INFORMATIONS ABOUT DISTANCE FUNCTIONS AND RELATED PARAMETERS

Each individual (or row) of a given database is here characterized by his covariates, so the distance between 2 individuals or groups of individuals depends on similarities between covariates
according to the distance function chosen by user (via the argument \code{dist.choice}). Actually four distance functions are implemented in \code{OT_joint} to take into account the most frequently encountered situation (see (3)):
\itemize{
\item the Manhattan distance ("M")
\item the Euclidean distance ("E")
\item the Gower distance for mixed data (see (4): "G")
\item the Hamming distance for binary data ("H")
}

Let two profiles of covariates \eqn{P_1} (\eqn{n_1} individuals) and \eqn{P_2} (\eqn{n_2} individuals) will be considered as neighbors if \eqn{dist(P_1,P_2) < prox.X \times max(dist(P_i,P_j))} where \eqn{prox.X} must be fixed by user (\eqn{i = 1,\dots,n_1} and \eqn{j = 1,\dots,n_2}). This choice is used in the calculation of the \code{JOINT} and \code{R_JOINT} algorithms.
In the same way, for a given profile of covariates \eqn{Pj}, an individual i will be considered as a neighbor of \eqn{Pj} if \eqn{dist(i,P_j) < prox.dist \times max(dist(i,P_j))} where \eqn{prox.dist} will be fixed by user.

For more details about the related algorithms integrated in \code{OT_joint}, please consult (2).
}
\examples{

### An example of JOINT model with:
#-----
# - A sample of the database tab_test
# - Y1 and Y2 are a 2 outcomes encoded in 2 different forms in DB 1 and 2:
#   4 levels for Y1 and 3 levels for Y2
# - n1 = n2 = 75
# - 2 discrete covariates X1 and X2 defined as ordinal
# - Distances estimated using the Manhattan function
# Predictions are assessed for Y1 in B only
#-----

data(tab_test)
tab_test2 = tab_test[c(1:75,5001:5075),1:5]

try1J = OT_joint(tab_test2, nominal = c(1,4:5), ordinal = c(2,3),
                 dist.choice = "M", which.DB = "B")

\dontrun{

### An example of R-JOINT model using the previous database,
### and keeping the same options excepted for:
#-----
# - The distances are estimated using the Gower function
# - Inclusion of an error term in the constraints on
    the marginals (relaxation term)
# Predictions are assessed for Y1 AND Y2 in A and B respectively
#-----

try1RJ = OT_joint(tab_test2, nominal = c(1,4:5), ordinal = c(2,3),
                  dist.choice = "G", maxrelax = 0.4,
                  which.DB = "BOTH")

### The same example of R-JOINT model as previously:
# - Adding a regularization term
# Predictions are assessed for Y1 AND Y2 in A and B respectively
#-----

try2RJ = OT_joint(tab_test2, nominal = c(1,4:5), ordinal = c(2,3),
                  dist.choice = "G", maxrelax = 0.4, lambda.reg = 0.9,
                  which.DB = "BOTH")


### Another example of JOINT model with:
#-----
# - A sample of the database simu_data
# - Y1 and Y2 are a 2 outcomes encoded in 2 different forms in DB A and B:
#   (3 levels for Y and 5 levels for Z)
# - n1 = n2 = 100
# - 2 categorical covariates: Gender and Dosage
# - 1 continuous covariate (Age) converted in ordered factor of 3 balanced
#   classes
# - Distances estimated using the Euclidean function
# Predictions are assessed for Y1 in B only
#-----

simu_data2 = simu_data[c(1:100,401:500),c(1:4,6,8)]

try2J = OT_joint(simu_data2, convert.num = 6, convert.clss = 3,
                nominal = c(1,4), ordinal = c(2:3,5),
                dist.choice = "E", percent.knn = 0.90,
                which.DB = "A")


### A last example of JOINT model with:
#-----
# - Another sample of the database simu_data using
# - n1 = n2 = 100
# - 3 covariates: Gender, Smoking and Age in a qualitative form
# - Complete Case study
# - The Hamming distance
# Predictions are assessed for Y1 AND Y2 in A and B respectively
#-----

simu_data2 = simu_data[c(1:100,401:500),c(1:4,7:8)]
simu_data3 = simu_data2[!is.na(simu_data2$Age),]

try3J = OT_joint(simu_data3, convert.num = 6, convert.clss = 3,
                 nominal = c(1,4:5), ordinal = 2:3,
                 dist.choice = "H", which.DB = "BOTH")

}
}
\references{
\enumerate{
\item Gares V, Dimeglio C, Guernec G, Fantin F, Lepage B, Korosok MR, savy N. (2019). On the use of optimal transportation theory to recode variables and application to database merging. The International Journal of Biostatistics.
Volume 16, Issue 1, 20180106, eISSN 1557-4679 | \url{https://doi.org/10.1515/ijb-2018-0106}
\item Gares V, Omer J (2020) Regularized optimal transport of covariates and outcomes in data recoding. Journal of the American Statistical Association, DOI: 10.1080/01621459.2020.1775615
\item Anderberg, M.R. (1973), Cluster analysis for applications, 359 pp., Academic Press, New York, NY, USA.
\item Gower J.C. (1971). A general coefficient of similarity and some of its properties. Biometrics, 27, 623–637.
}
}
\seealso{
\code{\link{merge_dbs}}, \code{\link{OT_outcome}}, \code{\link{proxim_dist}}, \code{\link{avg_dist_closest}}
}
\author{
Gregory Guernec, Valerie Gares, Jeremy Omer

\email{otrecod.pkg@gmail.com}
}
