% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_var.r
\name{select_var}
\alias{select_var}
\title{select_var}
\usage{
select_var(databa, Y, type_Y = "ORD", threshX = 0.9, threshY = 0.95,
  thresh_vif = 10)
}
\arguments{
\item{databa}{A data.frame containing variable of any type}

\item{Y}{A character string (with quotes) corresponding to the name of the target variable}

\item{type_Y}{A character string (with quotes) corresponding to the type of the target variable. "ORD" (Default) for ordinal type, "MUL" for multinomial type,
and "BIN" for binary type}

\item{threshX}{A double that corresponds to a threshold taken between 0 and 1.Beyond this threshold two variables are considered too highly correlated}

\item{threshY}{A double that corresponds to a threshold taken between 0 and 1.Beyond this threshold, a variable is considered too highly correlated with the target Y}

\item{thresh_vif}{A double that corresponds to a threshold (10 by default) of the Variance Important Factor beyond which a multicolinearity situation is highlighted}
}
\value{
A list of 6 elements:
        Y A simple reminder of the name of the selected target
        cor_Y A table of variables which are considered too highly correlated (Spearson coefficient) with the target
        VIF_PB 
        cor_X A table of variables which are considered too highly correlated (Spearson coefficient) with other covariates
}
\description{
This function preselect variables according to their associations with each other and according to their associations with a target variable
}
\examples{
require(Statmatch)
data(samp.A)   # Require Statmatch package
dat1 = samp.A[,-1]
sel_cov = select_var(dat1,Y = "c.neti")
}
